refactor(types): consolidate GameVersion and LoaderConfig in types.ts
Move GameVersion and LoaderConfig types from fabric.ts to types.ts to centralize type definitions and improve maintainability. Update imports in affected files to use the new location.
