+++ src/client/loaders/forge.ts
@@ -3,8 +3,8 @@ import { writeFile } from 'node:fs/promises';
 import { dirname, join } from 'node:path';
 import { downloadAsync } from '../core/download.ts';
 import { parseVersion } from '../handlers/version.ts';
+import type { GameVersion, LoaderConfig } from '../types.ts';
 import { fetchXmlWithRetry } from '../utils/fetch.ts';
-import type { FabricConfig, GameVersion } from './fabric.ts';
 import { getForgedWrapped } from './forgeWrapper.ts';
 
 type ForgeLoaderVersion = {
@@ -84,7 +84,7 @@ function getDownloadLink(loaderVersion: string) {
 	return `${baseUrl}/${loaderVersion}/forge-${loaderVersion}-${type}.jar`;
 }
 
-export async function setupForge(config: FabricConfig): Promise<string> {
+export async function setupForge(config: LoaderConfig): Promise<string> {
 	if (!(config.loaderVersion && config.gameVersion)) {
 		throw new Error('Missing version configuration');
 	}
